+++ circus-train-avro/src/main/java/com/hotels/bdp/circustrain/avro/transformation/AbstractAvroSerDeTransformation.java
@@ -15,6 +15,8 @@
  */
 package com.hotels.bdp.circustrain.avro.transformation;
 
+import static com.hotels.bdp.circustrain.avro.conf.AvroSerDeConfig.AVRO_SERDE_OPTIONS;
+import static com.hotels.bdp.circustrain.avro.conf.AvroSerDeConfig.BASE_URL;
 import static com.hotels.bdp.circustrain.avro.util.AvroStringUtils.argsPresent;
 
 import java.util.Collections;
@@ -22,20 +24,28 @@ import java.util.Map;
 
 import org.apache.commons.lang3.StringUtils;
 
+import com.hotels.bdp.circustrain.api.conf.TransformOptions;
 import com.hotels.bdp.circustrain.api.event.EventTableReplication;
 import com.hotels.bdp.circustrain.api.event.TableReplicationListener;
 import com.hotels.bdp.circustrain.avro.conf.AvroSerDeConfig;
 
 public abstract class AbstractAvroSerDeTransformation implements TableReplicationListener {
 
-  private final AvroSerDeConfig avroSerDeConfig;
+  private final AvroSerDeConfig avroSerDeConfig = new AvroSerDeConfig();
   private String eventId;
   private String tableLocation;
   private Map<String, Object> avroSerdeConfigOverride = Collections.emptyMap();
   static final String AVRO_SCHEMA_URL_PARAMETER = "avro.schema.url";
 
-  protected AbstractAvroSerDeTransformation(AvroSerDeConfig avroSerDeConfig) {
-    this.avroSerDeConfig = avroSerDeConfig;
+  protected AbstractAvroSerDeTransformation(TransformOptions transformOptions) {
+    if (transformOptions.getTransformOptions() == null) {
+      return;
+    }
+    Object avroSerDeOptionsObj = transformOptions.getTransformOptions().get(AVRO_SERDE_OPTIONS);
+    if (avroSerDeOptionsObj instanceof Map) {
+      Map<String, String> avroSerDeOptionsMap = (Map<String, String>) avroSerDeOptionsObj;
+      avroSerDeConfig.setBaseUrl(avroSerDeOptionsMap.get(BASE_URL));
+    }
   }
 
   protected String getEventId() {
@@ -51,7 +61,7 @@ public abstract class AbstractAvroSerDeTransformation implements TableReplicatio
   }
 
   protected String getAvroSchemaDestinationFolder() {
-    Object urlOverride = avroSerdeConfigOverride.get(AvroSerDeConfig.TABLE_REPLICATION_OVERRIDE_BASE_URL);
+    Object urlOverride = avroSerdeConfigOverride.get(BASE_URL);
     if (urlOverride != null && StringUtils.isNotBlank(urlOverride.toString())) {
       return urlOverride.toString();
     } else if (avroSerDeConfig.getBaseUrl() != null && StringUtils.isNotBlank(avroSerDeConfig.getBaseUrl())) {
@@ -68,7 +78,7 @@ public abstract class AbstractAvroSerDeTransformation implements TableReplicatio
     tableLocation = tableReplication.getReplicaTable().getTableLocation();
     avroSerdeConfigOverride = Collections.emptyMap();
     Map<String, Object> transformOptions = tableReplication.getTransformOptions();
-    Object avroSerDeOverride = transformOptions.get(AvroSerDeConfig.TABLE_REPLICATION_OVERRIDE_AVRO_SERDE_OPTIONS);
+    Object avroSerDeOverride = transformOptions.get(AVRO_SERDE_OPTIONS);
     if (avroSerDeOverride != null && avroSerDeOverride instanceof Map) {
       avroSerdeConfigOverride = (Map<String, Object>) avroSerDeOverride;
     }