+++ circus-train-avro/src/main/java/com/hotels/bdp/circustrain/avro/transformation/AvroSerDePartitionTransformation.java
@@ -24,8 +24,8 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 import com.hotels.bdp.circustrain.api.Modules;
+import com.hotels.bdp.circustrain.api.conf.TransformOptions;
 import com.hotels.bdp.circustrain.api.metadata.PartitionTransformation;
-import com.hotels.bdp.circustrain.avro.conf.AvroSerDeConfig;
 import com.hotels.bdp.circustrain.avro.hive.HiveObjectUtils;
 import com.hotels.bdp.circustrain.avro.util.SchemaCopier;
 
@@ -37,8 +37,8 @@ public class AvroSerDePartitionTransformation extends AbstractAvroSerDeTransform
   private final SchemaCopier copier;
 
   @Autowired
-  public AvroSerDePartitionTransformation(AvroSerDeConfig avroSerDeConfig, SchemaCopier copier) {
-    super(avroSerDeConfig);
+  public AvroSerDePartitionTransformation(TransformOptions transformOptions, SchemaCopier copier) {
+    super(transformOptions);
     this.copier = copier;
   }