+++ circus-train-avro/src/main/java/com/hotels/bdp/circustrain/avro/transformation/AvroSerDeTableTransformation.java
@@ -24,8 +24,8 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 import com.hotels.bdp.circustrain.api.Modules;
+import com.hotels.bdp.circustrain.api.conf.TransformOptions;
 import com.hotels.bdp.circustrain.api.metadata.TableTransformation;
-import com.hotels.bdp.circustrain.avro.conf.AvroSerDeConfig;
 import com.hotels.bdp.circustrain.avro.hive.HiveObjectUtils;
 import com.hotels.bdp.circustrain.avro.util.SchemaCopier;
 
@@ -36,8 +36,8 @@ public final class AvroSerDeTableTransformation extends AbstractAvroSerDeTransfo
   private final SchemaCopier copier;
 
   @Autowired
-  public AvroSerDeTableTransformation(AvroSerDeConfig avroSerDeConfig, SchemaCopier copier) {
-    super(avroSerDeConfig);
+  public AvroSerDeTableTransformation(TransformOptions transformOptions, SchemaCopier copier) {
+    super(transformOptions);
     this.copier = copier;
   }